re-introduce (now empty) translation for gluon-config-mode:pubkey
If the msgid is not found in the po file at all, the msgid itself
will be shown on the web page. This would lead to the following output:

"Your node's setup is now complete.
gluon-config-mode:pubkey

The node is currently rebooting and will try to connect to other nearby
Freifunk nodes afterwards. [...]"